+++ checkin/index.yaml
@@ -1,5 +1,9 @@
 indexes:
-
+- kind: MapUserToDeployment
+  properties:
+  - name: user_key
+  - name: deployment_key
+  
 # AUTOGENERATED
 
 # This index.yaml is automatically updated whenever the dev_appserver
