+++ checkin/views/main.html
@@ -79,7 +79,7 @@
                   <li><a href="/">Home</a></li>
                   <li><a href="sample">Sample Student Checkin</a></li>
                   <li>
-                    {% if user_link != None and user_link_text != None %}
+                    {% if user_link != None and user_link != 'None' and user_link_text != None %}
                       <a href="{{ user_link }}" target="_blank" >{{ user_link_text }}</a>
                     {% endif %}
                   </li>
@@ -88,9 +88,9 @@
                 <li class="dropdown">
                     <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Super Admin<span class="caret"></span></a>
                     <ul class="dropdown-menu">
-                      <li><a href="/deployments">Manage Deployments</a></li>
-                      <li><a href="/users">Manage Users</a></li>
-                      <li><a href="/reports">View Reports</a></li>
+                      <li><a href="deployments">Manage Deployments</a></li>
+                      <li><a href="users">Manage Users</a></li>
+                      <li><a href="reports">View Reports</a></li>
                     </ul>
                 </li>
               {% endif %}
@@ -98,7 +98,11 @@
               <li class="dropdown">
                   <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Admin<span class="caret"></span></a>
                   <ul class="dropdown-menu">
+                    {% if logo_url %}
+                    <li><a href="deployments">Manage Deployments</a></li>
+                    {% else %}
                     <li><a href="/deployments">Manage Deployments</a></li>
+                    {% endif %}
                     <li><a href="/users">Manage Users</a></li>
                     <li><a href="/visitors">Manage Visitors</a></li>
                     <li><a href="get_random_visitor">Run Raffle</a></li>
